app/requirements.txt
Django
psycopg2
Pillow
django-storages
boto
django-registration
